Sort: Use cub::WarpMergeSort for small sorts (n <= 128)
We currently use `bitonicSortKVInplace` for sorts of size `n <= 32`
which is a bit outdated since it doesn't use warp shuffles. It's also
unstable which forces stable sorts fall back to `radixSortKVInplace`
which is up to 4x slower in this small regime.

This PR adds a new kernel `warpMergeSortKVInplace` using
`cub::WarpMergeSort` to implement sorts with `n <= 128`. This results
in up to a 2x speedup for unstable sorts and up to 15x for stable
sorts, depending on the input geometry.

ghstack-source-id: 976f64f3bacf3c386241a81a187a86e33cea76c4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/96223